Fix ADIOS build issues.

+++ src/shared/adios_helpers_definitions.f90
@@ -251,7 +251,7 @@ subroutine define_adios_byte_scalar (adios_group, group_size_inc, &
   integer(kind=8),  intent(in)     :: adios_group
   character(len=*), intent(in)     :: name, path
   integer(kind=8),  intent(inout)  :: group_size_inc
-  byte,     intent(in)             :: var
+  integer(kind=1),  intent(in)     :: var
   ! Local Variables
   integer(kind=8)                  :: varid ! dummy variable, adios use var name
